+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Compatibility with TYPO3 v13
 
+### Changed
+- Default value of "ageOfDays" option in cleanup command lowered from 30 to 7 days
+
 ### Removed
 - Compatibility with TYPO3 v11
 - Widgets "Instance Starts" and "Instance Start Types"