+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.0] - 2024-10-01
+
 ### Added
 - Compatibility with TYPO3 v13
 
@@ -176,7 +178,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial pre-release
 
-[Unreleased]: https://github.com/jobrouter/typo3-process/compare/v3.0.1...HEAD
+[Unreleased]: https://github.com/jobrouter/typo3-process/compare/v4.0.0...HEAD
+[4.0.0]: https://github.com/jobrouter/typo3-process/compare/v3.0.1...v4.0.0
 [3.0.1]: https://github.com/jobrouter/typo3-process/compare/v3.0.0...v3.0.1
 [3.0.0]: https://github.com/jobrouter/typo3-process/compare/v2.0.0...v3.0.0
 [2.0.0]: https://github.com/jobrouter/typo3-process/compare/v1.2.0...v2.0.0